perf(api): cacheable GET variant for public indexer ops; split route modules
OPERATIONS/validation moved to operations.ts + variables.ts (pure module
split); new GET handler for an explicit CACHEABLE_OPS allowlist (five
zero-variable counts s-maxage=300 swr=86400; AllEndorsements pages
s-maxage=60 swr=600; no cache headers on upstream errors); CollectionsByUris
batch op added for the Ma Earth loader; allowlist + cache-header tests.
